backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser==2.2.0
Pillow==10.0.0
python-dotenv==1.0.0
reportlab==4.0.4